Corrected Drainage, Thermal Erosion working, usable KeyState to perform erosion, mouse scrolling, cleaning.

+++ Outerrain/Source/graph.cpp
@@ -4,7 +4,7 @@ static const weight_t max_weight = std::numeric_limits<double>::infinity();
 
 void ShortestPath::DijkstraComputePaths(vertex_t source, const adjacency_list_t &adjacency_list, std::vector<weight_t> &min_distance, std::vector<vertex_t> &previous)
 {
-	int n = adjacency_list.size();
+	size_t n = adjacency_list.size();
 	min_distance.clear();
 	min_distance.resize(n, max_weight);
 	min_distance[source] = 0;
